rename fps meter to game stats meter and add passing additional info about game state like game objects count and messages count
