+++ src/contrib/processorsPlugins/gameStatsMeterPlugin/index.js
@@ -1,0 +1,17 @@
+import ProcessorPlugin from 'engine/processorPlugin/processorPlugin';
+
+import { GameStatsMeter } from 'contrib/processors/gameStatsMeter';
+
+export class GameStatsMeterPlugin extends ProcessorPlugin {
+  async load(options) {
+    const {
+      frequency,
+      gameObjectObserver,
+    } = options;
+
+    return new GameStatsMeter({
+      frequency,
+      gameObjectObserver,
+    });
+  }
+}
